requirements.txt
asgiref==3.4.1
attrs==21.2.0
Django==3.2.7
django-cors-headers==3.10.0
django-filter==21.1
django-tinymce==3.3.0
djangorestframework==3.12.4
drf-spectacular==0.20.2
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.1.2
Pillow==8.4.0
psycopg2-binary==2.9.1
pyrsistent==0.18.0
pytz==2021.1
PyYAML==6.0
sqlparse==0.4.1
uritemplate==4.1.1
